+++ field.py
@@ -2,10 +2,9 @@ class Field:
     """
     Abstract base class for a field in a message.
     """
-    def __init__(self, name, length, optional=False):
+    def __init__(self, name, length):
         self.name = name
         self.length = length # length in bits
-        self.optional = optional
 
     def decode(self, data):
         """
@@ -18,7 +17,8 @@ class Field:
     def encode(self, value):
         """
         Convert value to self.length bits of data (in an LSB-aligned bytes object) to
-        build a message. Raise a ValueError with an appropiate message if this is not possible.
+        build a message. Returns a tuple of (data, len_of_data_in_bits). Raise a ValueError with
+        an appropiate message if this is not possible.
         """
         raise NotImplementedError
